MB-65328: Upgrade any CA cert with missing subject attributes
In 7.6.0, support for the 'OU' attribute type was added for formatting
certificate subjects. However, existing certificates are not upgraded
to include this attribute, which can cause issues for consumers of the
certificates as returned by our REST APIs, since they won't include the
full subject as originally specified.

On Morpheus upgrade, we will update the subject field, so that it
includes all currently supported attributes, specifically 'OU'.

Since this field is stored alongside the pem, there is no change to the
pem itself.

Change-Id: Ia556ef566258659121b8a8eacc300dd40e194509
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226535
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>