MB-66321: Reduce logging around dropping app telemetry clients
When dropping a client, this causes the associated process to be
terminated, which then causes additional logging when that DOWN message
is handled. We should avoid this additional logging when it's not
needed, as follows.

Don't drop the connection when failing to call GET_TELEMETRY if the
failure reason was 'dropped', since the connection will certainly
already be dropped. Also log at debug level instead of warning.

Make sure to demonitor the process when we drop it from our client map,
so that we don't get a DOWN message that would need ignoring.

If we get a DOWN message from a pid we don't recognise, then log this
at warning level, but don't try to drop it.

If we get a DOWN message from a pid which is still associated with a
connection, then we still wish to log, since it will likely have been
terminated somewhere which doesn't log the cause. After logging, we'll
still drop the connection.

Change-Id: Idd5f3029ed17747d672e94dcdd41c8791941edac
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226472
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>