cluster_tests: Fix using existing cluster
When using an existing cluster, the is_node_started function crashes,
as there are no processes associated with the cluster.
Also, the cluster.destroy() function is still called for existing
clusters, even though it now asserts that it is not called with an
existing cluster.

Fix these two things, and also replace the cluster.existing_cluster
field with a helper function, since it is rarely used and is equivalent
to `cluster.start_args is None`, which is used much more frequently.

Change-Id: I756836f9c0b6faea5347bea2d76e77db9c10519d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226473
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>