MB-66153: fix collecting FFDC vitals files ...
... that don't end with a newline. The core issue is that we weren't
clearing the wraparound on the DoubleStream which causes us to not have
the inner RedactStream get it's clear_wraparound called. There were also
conflicts with how we normally just "passthrough" binary files. Because
we cannot do that with these, another mechanism had to be added. Added
"skip_extension_check" to the RedactStream class so it can ignore that
feature for the FFDC collection task.

Change-Id: I8a79663426b63bf412732e9235c65f4340b13577
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226497
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>