MB-66153: [BP] fix collecting FFDC vitals files...
... that don't end with a newline. The core issue is that we weren't
clearing the wraparound on the DoubleStream which causes us to not have
the inner RedactStream get it's clear_wraparound called. There were also
conflicts with how we normally just "passthrough" binary files. Because
we cannot do that with these, another mechanism had to be added. Added
"skip_extension_check" to the RedactStream class so it can ignore that
feature for the FFDC collection task.

Change-Id: I5ac4d38c5e6c8807220cec5f711a8e674d024d39
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227306
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>