MB-62886: Clean up residual rebalance state
If a rebalance completes but the ns_orchestrator has crashed then the
rebalancer pid doesn't get cleared. As a result we mistakenly, at least
in some code paths, believe a rebalance is running.

Prior to this change this condition was detected (and fixed if needed)
during bucket janitoring. However if the cluster had no buckets
configured bucket janitoring wouldn't occur and the condition would not
get fixed. With this change the checking (and fixing if needed) is
separated out of the bucket janitoring so it happens even if there are
no buckets configured in the cluster.

Change-Id: Ib268b7f94a2f5889e6ff4b070571c1f144418f88
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226606
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>