[cluster tests] Fix expectations in dont_remove_active_dek_test
We can't really expect that with rotation interval set to 0
we will have 1 DEK on disk.
If reencryption of configs takes about a second, we will almost
always have 2 DEKs on disk.
To fix the problem we should disable 1 second rotation, and
only then wait for all but one DEK to be removed.

Change-Id: I5a40d3d66b70d9331a8bdbf715b5d3b979b7eed8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226807
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>