MB-66415: Add "LowUrgency" tag for certain tests
The cluster_tests CV job now takes an unreasonably long amount of time,
sometimes as long as 2 hours. A large source of this is a small
minority of test involving rebalances and other slow operations, which
are not all necessary to run on every patchset.

To allow us to filter out these tests, and run them instead in a
post-commit job, we need to give them a tag. While the obvious tag
would be something like "post_commit", since it makes sense to be
explicit about what the tag is for, that doesn't give us the
flexibility to choose the job depending on the branch. Naming things is
of course a hard problem, but it's better that we have a bad name than
continue to have an almost 2hr job. We are calling the tag
"LowUrgency", since these tests don't need to be fixed urgently, and
so they are safe not to run in CV on every patchset. To make it clear
that these tests may or may not run depending on the branch, this is
documented alongside the tag decorator function.

We can rename this if we come up with a better name (and just have
both tags in the jenkins job configurations during the transition).

The tests which have initially been giving these tags were those
identified as having >10s runtime, excluding a small number of
particularly valuable tests. We can always remove the tag from these
if we identify any with a greater impact than realised, and add more
as needed.

Change-Id: Ib9507bb4a7da80619305827111c3b644be5c7c3f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226843
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>