MB-66073: Add x86_64 micro arch detection to startup script
Ensure that the server does not start up if we can correctly detect that
the (x86_64) CPU does not support x86_64 v3 or newer microarchitecture.

The logic for the check is contained in a new file,
validate-cpu-microarchitecture.sh. This is separate from the startup
script itself so that it may also be included in its entirety in the
rpm/deb pre-installation scriptlets.

The environment variable COUCHBASE_DO_NOT_VALIDATE_CPU_MICROARCHITECTURE
can be set to any value to bypass the new check. If running Couchbase
Server as a service, this can be set in a systemctl override as such:

```
$ sudo systemctl edit couchbase-server
[Service]
Environment="COUCHBASE_DO_NOT_VALIDATE_CPU_MICROARCHITECTURE=1"
```

This option has been added as an environment variable as it only there
as a safety valve, in case a customer runs into a situation where the
check itself causes problems.

Change-Id: Ifd7a3dc177615ae983c747581073e0a164c3cf3e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/226947
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>