MB-61292: Add /controller/forceEncryptAtRest/*
This call is supposed to make sure that corresponding data
is fully encrypted at rest.
For instance, if one enables encryption for a bucket, it is likely
that only data that was written after the encryption was enabled
will be encrypted.
In this case, the following call
POST /controller/forceEncryptAtRest/bucket/<bucket_name>
will initiate encryption for all unencrypted data in the bucket.
The data that is already encrypted will not be re-encrypted.

It works the same for logs, audit, and config:
POST /controller/forceEncryptAtRest/logs
POST /controller/forceEncryptAtRest/audit
POST /controller/forceEncryptAtRest/config

Change-Id: I2a6cbe6fd73896759e4b8436274f7bd37dae0f1f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227151
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>