MB-64582: Refresh for in use deks in ale disk sinks
Currently the ale disk sinks referesh the in use deks anytime a file
is rotated or if dek drop is attempted. It is possible a user removed
some files manually in which case we should also refresh in_use to allow
any garbage collection of unused deks. Infrequent interval of once in
24 hours was picked as it doesn't seem too useful to do it more
frequently because if deks ever need to be dropped, it would referesh
the in use from disk at that time in that case anyways.

Change-Id: Ieba41dbb18ac2ec4f0c5dc6d955a98e420bbb8a0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227198
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>