MB-66268: Fix overflow issue with timeout value
The underlying type of time.duration in go is a int64 type. It keeps
time in ns units, so when converting large values from ms unit to
duration, it can overflow the duration type of int64. It doesn't make
much sense to allow really large timeouts to begin with, so with this
change we cap the timeout to 5 minutes on input and also internally on
client side.

Change-Id: I90298aa903a600928f6f433f0d8ef78b644b4bce
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227208
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>