MB-66433: Initialise disconnected nodes with correct afamily
Disconnected nodes weren't being initialised with the same address
family as the connected ones, meaning that the requirement wasn't
fully implemented.

That meant that NodeAdditionWithCertsIPv6Tests fails when run alone,
because it was previously depending on other tests first running
and coincidentally initialising the disconnected node as ipv6.

Also updates the hostname, such that disconnected nodes have the same
hostname, whether they have been part of a cluster or not.

Change-Id: I78047604d687f64cb32623f0e01cd8610f9fd3b4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227351
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>