MB-51771 - permit client certs for new node authentication
When a new node is added to a cluster, if certificates are used it is
not necessary to ask for a username/password.

New nodes can be added in two places:
- In the Add Node dialog of the existing cluster
- In the Join Cluster dialog of the new node

In both places we add radio buttons where the user can choose between
login/password and certificate authentication. If certificate
authentication is selected, the username and password are hidden, and
the server request has the `clientCertAuth` parameter set to true.

Change-Id: Ia484b62acaad368d9829cc106c0038213cce4ba2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227436
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>