MB-61292: Don't use funs in calls between nodes
Calls between nodes should not use funs as node may run different
versions of the code, so the funs will crash.

This change replaces the funs with {M,F,A}.

Note that we also can't perform the is_writable check not on the
node that is handling the http request (because req and #auth_res
format may change between versions).
For this reason the MFA calls is_writable on the node that is
handling the request.

Change-Id: Ia109f420fecbedb140f2a034239d17b01330ba99
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227514
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>