MB-65435: server groups cluster tests ...
... greedy vbmap tests. Largely focuses on constructing poorly balanced
server-groups s/t we are more likely to generate an initial map outside
of the predefined replica range of 5. This is also why we need to use
the full number of vbuckets, since it's almost impossible to trigger the
issue when using so few vbuckets like a lot of the other tests do to
save time.

Change-Id: I2c980c8cfff01f80933878dd91e6a1343e3e3a8e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228184
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>