MB-66178: Adjust dump-bucket-deks and dump-keys utilities...
... to bucket directory name change

We can't support --bucket parameter anymore.
Replace it with --bucket-uuid, but it doesn't seem to be
very convenient as the bucket UUID is not always available.
For that reason, it is made possible to search for keys
in all bucket directories. Since the key name is unique,
the bucket id is not strictly necessary to find the key.
Specifying bucket id can speed up the search though.

Interface changes:

- dump-bucket-deks:
    The --bucket parameter is replaced with --bucket-uuid.
    Note that --bucket-uuid is not mandatory. If it is not specified,
    all buckets are searched for the key.

    Also the --bucket parameter is ignored for now if passed to let
    users migrate their scripts.

- dump-keys:
    The (--key-kind or --key-dir) parameter group made optional.
    If neigher of the two is specified, all key kinds are considered
    when searching for keys. Specifying the --key-dir or
    --key-kind limits the number of directories searched for keys.

Change-Id: Ia3211f9bff91a25879f0ddb8ca62e9f977a43bbc
Reviewed-on: https://review.couchbase.org/c/ns_server/+/227580
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>