MB-61888: Handle retried graceful failover with expected topology
Stash the opts in the rebalancer state such that we can re-check them
on the retry. The topology should not materially change, the
orchestrator should not attempt the retry if it has, but it could be
the case that we failed before checking the topology so we should make
the check again on retry.

Change-Id: Iaf81e9a6a02aabefa408c5b95258a5f485930f76
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228047
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>