MB-51771 - set certificate auth default based on current node cert
When adding a new node to the cluster, the UI should default to
certificate-based authentication only if the current node has an
uploaded client certificate.

This change works for both 'Add Server' from an existing cluster, and
'Join Cluster' from a newly created node.

In the case of 'Add Server' it needs to get the list of client certs
and find the one that matches the current node name.

In the case of 'Join Cluster' from a newly created node it needs only
take the first certificate since there is only one node.

Change-Id: Iabc6faa7820a7d146ebf3e17848e63b419ea8826
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228072
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Pavel Blagodov <stochmail@gmail.com>
Well-Formed: Build Bot <build@couchbase.com>