MB-66860: use threads to fix deadlock in cbcollect
It turns out, it is possible to cause deadlocks when reading both stderr
and stdout with fixed-size buffers. This deadlock potential is
documented and the suggested fix is to utilize Suprocess.communicate
because it will safely get both the stdout and stderr streams of data
without the risk of this deadlock...

That will, however, buffer the entire output of the
process (stdout/stderr) in memory. This could be problematic for
processes that produce a lot of output. Instead, this patch adapts
suprocess.communicate for our use-case s/t we don't buffer everything
before flushing to redactors and then zipfile.

This is done using threads. There are a few other ways this _could_ be
accomplished but considering this is the way python does it internally
for suprocess.communicate -- it seemed like the way to go.

Change-Id: Id04b631cded746055da96fbf052a3ec163eaf9d0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228700
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>