MB-67026: Exclude client cert for cbimport with non-ssl port
When the ssl port is not required, we were still sending the client
cert env variables to cbimport, which causes it to give an error as it
doesn't expect these values to be populated when using the non-ssl
port.

We should only provide client certs when we are using the ssl port and
client cert auth is mandatory.

Change-Id: I039bb4f9a80f582e3cb49be88e5f11cee4eb7e35
Reviewed-on: https://review.couchbase.org/c/ns_server/+/228800
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>