MB-67857[BP]: Fix fake_chronicle_kv
The changes to fake_chronicle_kv and failover_tests needed to be
backported to avoid conflicts in subsequent backport commits. Other
changes to code not in trinity could not be backported, so this commit
is quite different from the original.

The test code around config_profile is also quite different in trinity,
so the changes to failover_tests in the original commit could not be
made directly, and had to be roughly approximated with the code already
existing in trinity.

Change-Id: Ib45963cc4d8d8098841bccdf5c198dcd60f8f21a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232191
Well-Formed: Restriction Checker
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>