MB-62644: Validate cert chains in engage_cluster
When adding a node to the cluster, connectivity checks are performed
using the trust store of the node being added. When the node gets added
in, and joins the cluster's chronicle, it will then have the cluster's
trust store, potentially causing connectivity to be lost.

For instance, consider the case where the node being added has an
intermediate cert in its trust store, which forms part of the chain for
the node and/or client cert of that node, but which is not in the trust
store of the cluster. The TLS handshakes may succeed, since the
intermediate cert in the trust store will get included in the
node/client chain, by Erlang's SSL implementation. However, after the
trust store gets overridden, the intermediate cert would no longer be
available to Erlang for providing in the handshake, so the node/client
cert chain may no longer be verifiable by other nodes in the cluster,
as they will be missing that part of the chain.

To ensure that this situation can't occur, we should validate the cert
chains using the cluster's trusted certs, and reject adding the node if
it fails.

Change-Id: I63a3903ba456c477dd08f5fe996c690586313152
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229190
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>