Use a better balanced vb map in failover tests
Currently the map can be unbalanced as it is random. This causes issues
in tests as they can sometimes fail over a node that would not be
failover-able were the map balanced. Use a balanced map in the tests
as we do in the Server to address this.

If this continues to be an issue, we should just fix the maps in each
test rather than attempt to generate them (for convenience).

Change-Id: I1e3e12ac021bdc6d433a6292363b5b02c9e8f328

Reviewed-on: https://review.couchbase.org/c/ns_server/+/229354
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>