MB-61292: Check keys before applying them
In some cases it doesn't make sense to apply keys if we could not
read some of them.
For example, we should not push keys to memcached if active key
is not available.
It also seems very dangerous to apply snapshot with missing keys
for chronicle encryption, because any encryption error in chronicle
will lead to node failure, so it is better to wait until we can
get all the keys.

Note that in some cases the explicit check is not needed, because
the error will happen anyway when we try to encrypt data.

Change-Id: I17a5a005fffa17bc8e82b68668f0a7960822636b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229447
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>