MB-61292: Retry create bucket if key is not available
If ns_memcached crashes, the whole supervisor tree is restarted,
which leads to log pollution and high CPU usage.

This change tries to address it, and retry after timeout gracefully,
because the "key not available" error is something that can happen
during normal operation. For instance, a DEK can be unavailable
when there is no connectivity to external key server (e.g. KMIP),
so DEK can't be decrypted.

Change-Id: I68ad98c423b1614c978a9b1dd818b07952daa362
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229448
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>