MB-66812: Switch metakv rev to opaque SHA256 revision format
This change addresses the issue where vector clock revisions were too
long. Switch to the new format:
- Use a SHA256 hash of the vector clock

This does not change anything about the metakv vector clocks internally.
Services use `rev` to perform a CAS operation. `rev` is queried on a
node and passed on subsequent mutate operations - to ensure that the
underlying value is updated atomically. During a mutate operation, `rev`
is computed afresh and compared (strict equality) with `rev` in the
incoming request. If they differ, the operation is aborted.

This does not require any changes in the API nor to services because:
- Services treat `rev` as an opaque identifier. `rev` queried during a
  Get is passed as-is during subsequent mutate operations.
- Services use `rev` on a node-local basis. There is no need to
  consider distinct `rev` formats on nodes running different versions.
- During an upgrade, a service is restarted along with ns_server/cbauth.
  i.e. a node running a particular version always uses the same format
  for metakv Get/Set/Delete.

There is no need to account for mixed `rev` formats.

Change-Id: I79998ef844163a06858f2fd842cd25b6a65405f1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229982
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Peter Searby <peter.searby@couchbase.com>