MB-67857[BP]: Pull failover test mocks into new module
We can re-use much of the failover_test mock setup in future rebalance
tests. To do so, we will move much of the mock setup to a new helper
file and introduce a new pattern to use these helpers.

The mock helpers will set up the processes or mocks that various modules
set up. These need to be torn down later. As such, we must track these
artifacts. To do so, we will track them in a map (generally referred to
as the PidMap) which stores a pid, or some other marker to determine
that this module has been started/mocked. The setup helper functions can
be used to set up a list of modules, skipping any duplicate setup
required. The map can be passed to the shutdown_processes/1 function to
shut down the set up processes. Mocks can be unloaded wholly with
meck:unload/0.

I'm unsure if this an ideal longer-term solution/pattern, but it solves
the immediately problem of mock re-use and we can consider changes as we
use it more.

Change-Id: I841682f91e055621afaaa149be969cb12508ae52
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231994
Well-Formed: Restriction Checker
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>