MB-67345: [BP] Intermittent failure - derived_stats_test
This is a backport of MB-62347.

The couch_docs_actual_disk_size stat is derived from
kv_ep_db_file_size_bytes and so there's a window where the stat is
not yet available as the base stat hasn't been emitted. The test
now waits for the availability of the stat.

Change-Id: Ia80c4e1b0b5b201cab66320482d130a6fbba136d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/229906
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>