MB-61292: Recalculate DEKs info after bucket creation
cb_cluster_secrets tries to recalculate it automatically when it
creates DEKs for just created bucket (created in config), but it
is racy, because ns_memcached is being started in parallel (bucket
is not necessarily created in memcached yet). If the bucket doesn't
exist in memcached when DEK info is being recalculated, the bucket
data encryption status will be set to "unknown", and will be
recalculated only in 10 minutes (by timer). This behavior is very
confusing from user point of view, so to avoid it, recalculation is
forced right after bucket creation (when data status can be retrieved
for sure).

Change-Id: I3bac74603379fa241cf9574d0951f5404f63bc46
Reviewed-on: https://review.couchbase.org/c/ns_server/+/230992
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>