cluster_tests: Exclude n1ql/index from serverless sample test
There is a bug with n1ql/index loading sample buckets in serverless
profile - MB-66431.

Since serverless is not in development, we should just ensure that
the requirements for this testset do not permit this case to be hit,
which is done most simply by excluding these services.

Change-Id: Ib98c5d6f2552b1e74cda36f13983217c1f34df18
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231149
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>