MB-66820: Refactor rebalance checks for re-use
This change does two things:

1) Moves the rebalance param map building into a new function
2) Moves the rebalance safety checks into a new function

This is done such that we can re-run parameter building and checks
in the leader activity on a consistent config in a future change.

Two notable changes have been made to do so:

a) The rebalance retry check parameter building was separated from the
check itself - this is such that the parameter building function does
only that.
b) The fusion rebalance plan check parameter building was similarly
separated from the check.

Change-Id: I385cf89baf49ad4ff3a97f66efdae05094bfcf56
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231217
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>