MB-66765: Disable memcached crash dumps during log/data encryption
When log/bucket data encryption is enabled on the node, we ensure
that breakpad is disabled in memcached config. This is done by using
the encryption state of the node to determine the breakpad value
for the memcached config.

It is important to allow the user to change the breakpad value to
allow debugging even when encryption is enabled. For that, a new
global cluster wide setting has been provided to force the breakpad
to enabled.

Change-Id: I8fa85d2e1e4bbad988aab29251e50dbe38a09ed9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231406
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>