Put fake_chronicle_kv seqno in snapshot
Such that we can compare and swap all values in a future change, which
will allow us to support concurrent writes in fake_chronicle_kv.

Notably, the chronicle_metakv tests verify the state of the entire
chronicle snapshot. As such, remove the seqno from the snapshot on
return such that those tests can continue to do so.

Change-Id: I9611e5639a43e34ab3caacd567684b0fb2f38339

Reviewed-on: https://review.couchbase.org/c/ns_server/+/231615
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Peter Searby <peter.searby@couchbase.com>