MB-67804: Make sure cb_dist:select does not block net_kernel
All erlang proto modules try to resolve the hostname
when select() is called. This can block the net_kernel and lead
to node start problem (while we are waiting for dns couchdb
timeout expires).
At the same time, the name resolution is not really needed in
cb_dist:select() as it already knows the protocol (afamily and
encryption) that it will use.

Change-Id: If0c7af5298ed04c872165b4efbf6e7d7a8c21ad4
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231603
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Peter Searby <peter.searby@couchbase.com>