MB-67857: Avoid menelaus_cbauth_worker message queue growing
Because we synchronously notify cbauth instances of creds info, and
the creds info is generated outside the worker process, we can end up
with a queue of updates filling up the menelaus_cbauth_worker process.

The obvious solution would be to simply generate the creds info within
the worker process, such that we could instead flush all notify
messages each time we process one, avoiding the message queue backing
up. The issue with this is that it would mean duplicating the
management of creds info in each worker process, which is currently
handled in just menelaus_cbauth.

Instead, it is simpler just to flush the notify messages when we
process one, while retaining the info from the last message.
This avoids the message queue backing up, ensuring that each time we
notify cbauth, it is with the latest info that the process received.

Change-Id: I3c3b15dd7b81f005524a24d0614895a6e8fcb79f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/231636
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Peter Searby <peter.searby@couchbase.com>