MB-67138: Fix intermittent node addition failure
In a couple of node addition tests we change the security settings of a
node to require 'mandatory' client certs. As a result of this the web
services on the node are restarted. The test "waits for the web server"
by sending a /pools request. The problem is the request could be sent
before the web server restart...and as a result a subsequent REST call
to that node will get a "connect_closed" error.

To handle this we check for the "connect_closed" error and then "wait
for the web server" and then try the REST call we want to do.

Change-Id: I591557b3f632f9376dce9785fc708c061b84c37a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232731
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>