MB-68060 Edit Storage Backend.
Storage Backend may be edited after the bucket is created.
Buckets and Servers are entering a special pending state after this - storage backend pending and/or eviction policy pending; the statuses are shown in the bucket list and server list. The pending state should go away after these actions are made on all the servers that contain the pending buckets:

If only eviction policy is pending:
- For each Data node, Swap rebalance or Graceful failover & delta recovery and rebalance to complete the process of changing the eviction policy.
If storage backend is pending:
- For each Data node, Swap rebalance or Graceful failover & full recovery and rebalance to complete the process of changing the storage backend.
When a bucket is in a pending state due to changing storage backend and/or eviction policy there is a restriction on what fields can be edited until the pending state is gone - so a lot of the fields are disabled in this case (the only fields that can be changed are: 'storageBackend', 'numVBuckets', 'evictionPolicy', 'noRestart', 'ramQuotaMB').

Under Eviction Policy there is a new checkbox that sends a new property to the bucket endpoints: "noRestart" - checking this allows the eviction policy of the bucket to be edited but not restart the bucket; instead the bucket goes into the eviction policy pending state.

Change-Id: I70422e33ba60f17103266eef27ffa107c0100ecb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232430
Tested-by: Eben Haber <eben@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Eben Haber <eben@couchbase.com>