MB-67857: Ensure cbauth_many_notify_t actually works
The test doesn't actually fail when the assertion is reached, as the
meck:wait can return successfully even when the function body fails.

To avoid this, just send the Port to the test process, for it to run
the assertion directly.

Change-Id: Iee47e4d2052a9214340ada9a00544dce07943e4f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232542
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Restriction Checker
Well-Formed: Build Bot <build@couchbase.com>