MB-68168: Handle exdev rename error when retiring keys
It may happen that data dir and retired keys dir are located on
different filesystems. In this case, the rename operation will
fail with an exdev error, so we should fall back to copy+delete
the file instead of renaming it.

Key files are small so the performance impact of this fallback
should be negligible.

Change-Id: I26a71c4aaa4e65d9a098b167276835708a4bce31
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232659
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>