MB-68227: Fix /forceEncryptionAtRest endpoint...
... when called while encryption is disabled

This endpoint normally is supposed to just set
force_encryption_date in config, but this value doesn't matter
when encryption is disabled.

Forcing decryption when encryption is off basically means
getting rid of all DEKs, which is exactly what dek_drop_datetime
is for. Hence the fix: if encryption is disabled, just
set dek_drop_datetime to the same value as force_encryption_date.

Change-Id: I1369f874ba71df4616da64002b83faebc6513097
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232739
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>