Change cache in gosecrets to respect proof in keys
Currently, the code assumes that keys in cache have passed the
integrity check, while it is not always true (because we are not
always able to check it, for instance we can't do that early
during start).
This change makes sure we save the proof information in cache.
If later we want to use it, we have to check proof first.

Change-Id: I1faf590b17bb2cfd2f818ee155d97002be639c93
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232818
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>