MB-68293: Don't use cached keys when testing keys
When testing keys the goal is to test that the key is available,
which includes the fact that keys that encrypt it also available.
Basically, we want to test availability of the whole chain of keys,
up to the root key.
If at least one key in the chain is cached, we cannot be sure that
the rest of the keys in the chain are available, so we should not
use cached keys when testing keys.

Change-Id: Ia3a6e21f8e40ff96571c6c0cd8772901ac604bc1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/232986
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>