MB-68263: Increase retries when ejected oneself
A node ejecting itself from the cluster involves telling another node to
"shun" us. This involves chronicle and ns_config operations and
currently must be completed in 250ms * 10 retries. This deadline was set
way back (only checked back to 3.0.1), chronicle didn't exist, and it
seems reasonable to increase the number of retries from 10 to 20. This
effectively allows 250ms * 20 retries for the eject to complete.

Change-Id: I0540bf024f1f4d5c3aca8881a4976ab9fb419ece
Reviewed-on: https://review.couchbase.org/c/ns_server/+/233465
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>