MB-68456: Make full eviction default bucket magma policy
If magma is not specified as the storage backend, then the default
policy will remain value eviction. This happens because is_magma/4
defaults to assuming that the bucket is couchstore, which is no longer
the case.

Update the function to default to magma, and correct the tests that
break accordingly.

Change-Id: I7ad891f897486f691348b61a46133aa462912f72

Reviewed-on: https://review.couchbase.org/c/ns_server/+/233573
Reviewed-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>