MB-64812: [FBR] Tidy up rebalance test teardowns
Add a base class teardown call to KVFileBasedRebalanceEncryptionTests.

Move the bucket deletion to teardown rather than test_teardown, as it
is created in setup. We must also delete buckets before we delete
encryption keys in KVFileBasedRebalanceEncryptionTests, as they
cannot be deleted if the bucket is still in existence.

Change-Id: I4138e138f6eb2755acbcdb0459f32a7041bdcb47

Reviewed-on: https://review.couchbase.org/c/ns_server/+/234385
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>