MB-61655: Do not pass access log path to ephemeral buckets
KV only allowed this due to incorrect validation of parameters with
bucket type requirements. The issue applies only to the initial bucket
config string. The parameter was always rejected on ephemeral buckets
when set via SetParameter (if you try to set it after the bucket is
created).

Change-Id: I51fec256883b91e0f0fdd972110579b455e8adaf
Reviewed-on: https://review.couchbase.org/c/ns_server/+/234710
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>