MB-68185: Fix FilteringEnabled criteria
Caught by dialyzer in OTP 28.

FilteringEnabled was checking for EventStates being `[]`, when it is
wrapped with `{...}`, so that case is impossible.

This meant that when no audit events were configured enabled/disabled
differently from default and there were no disabled users, the
filtering was incorrectly enabled, even though it was unnecessary.

Change-Id: I1fa8bc3a15aa55a31e8486040009334ea7fda948
Reviewed-on: https://review.couchbase.org/c/ns_server/+/234664
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>