MB-61655: Add CMD_VALIDATE_BUCKET_CONFIG support
Added support for the ValidateBucketConfig MCBP command, which can
validate a bucket config, before creating/editing a bucket.

The Buckets API now uses that to perform additional validation on the
parameters.

This is groundwork for allowing parameters to be validated via the Data
Service. For now, all this patch does is validate via memcached and
report errors. Note that API parameters are passed down as-is, without
any name mapping.

No extra bucket parameters are passed down to memcached yet.

Change-Id: I3433d727c5539a7fa66c40d0ae4128836fd13c66
Reviewed-on: https://review.couchbase.org/c/ns_server/+/235291
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>