MB-61655: Update BucketConfig with extra params validated by KV
Validated parameters which are not recognised by us, but are recognised
by KV are stored in the BucketConfig, under extra_params.

They are returned by the Buckets REST API, and are merged into the
bucket info.

In later patches, these extra params will actually be used to
create/update the bucket in KV.

Currently the parameters must be specified in the KV format, snake_case.
In a future change, we will be able to specify them in camelCase similar
to other parameters supported by the REST API.

Change-Id: Icb0e9d8231d16d68bbe4ddd7eac5f20e1ee353a2

Reviewed-on: https://review.couchbase.org/c/ns_server/+/235292
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>