MB-61655: Pass down the extra_params from BucketConfig
The extra_params are parameters we have previously validated in KV. We
can merge those in the bucket config string when creating a bucket. We
can also dynamically set them in ensure_bucket, via SetParameter.

This patch implements the above, and adds support for
SetParameter(config), which is a new reload type which can be used to
set any parameter.

Additionally, when ensure_bucket detects an extra_param has changed, it
will do a ValidateBucketConfig call, in order to determine whether the
parameter can be changed at runtime or whether it requires a restart of
the bucket. This is implemented by populating the restart atom in the
parameter props list, before we proceed with the ensure_bucket check.

Change-Id: I64dc710e35f5e4359e4ee660cd7767bbb845dfca
Reviewed-on: https://review.couchbase.org/c/ns_server/+/235293
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>