MB-61655: Fix up handling of numbers in ensure_bucket
Now that we can validate parameters via KV and store them in the
BucketConfig, we quickly run in to a problem.

Numeric parameters do not compare equal with our current logic, so
ensure_bucket resets them in every run. This appears to be happening,
because:

1. When we validate via memcached, it returns the value in JSON
2. The json parser parses numbers out as floats in Erlang
3. Our code formats floats with decimal point ("5.0", not "5").
4. Memcached STATS returns the number without decimal ("5").
5. We use string comparison to determine if the config has changed

And we always compare "5.0" (which is float in the bucket config because
it was parsed from the JSON response from validation), and "5" which is
how memcached will output the number. This concerns even entirely
integral parameters.

This patch compares float values ending in ".0" with their truncated
integer values. If they are equal, the number is formatted as an
integer. Fractional values remain unchanged.

Change-Id: I24697aaaa5c67981ac09d388053152dff2c022e7
Reviewed-on: https://review.couchbase.org/c/ns_server/+/235294
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>