MB-67396: Refactor to have AWS key callbacks use common code
The reason for this is because AWS, GCP, and Azure support logic
for most key callbacks ends up being identical, so this change
makes it such that the common code lives in the common file.
This will allow AWS, GCP, and Azure to use the same logic instead of
having to create identical seperate files for duplicate logic.

Change-Id: I35f7bc5a3bd6933ccfaceaeec8945842df87be7d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/235403
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>