MB-68505: Correctly set uploaders for trivial moves
It can happen that there's no need to move the data but
uploader still needs to be assigned for given vbucket. Update
uploaders map with such uploaders right in the beginning of
the rebalance and rely on janitor to start actual uploaders
later on.

Test added here creates rebalance with trivial moves by failing
over the second node, so the vbucket map goes through the following
transformations:

initial
vbmap = [A, A, B, B], uploaders [A, A, B, B]
failover
vbmap = [A, A, A, A], uploaders [A, A, undefined, undefined]
rebalance out
vbmap = [A, A, A, A], uploaders [A, A, A, A]

so all 4 moves are trivial and uploaders on vb3, vb4 still need
to be set.

Change-Id: If9cf7000ee4369661d03e495dab74136e37262bc
Reviewed-on: https://review.couchbase.org/c/ns_server/+/235891
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>