Do not fall into infinite loop if wait for uploaders to stop...
timed out

Instead remove the namespace from the work queue and rely on
the next check to find that it is still not deleted. This gives
a chance to delete other namespaces that should be deleted
while uploaders for this one are stopping.

As a side effect it fixes crash caused by the attempt to print
the result of delete_data (which is not string) as ~s

Change-Id: I3e810a5148abaa09e53570bfb3e019834bd796e5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/236067
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>