MB-69210: Treat missing vbuckets in memcached "uploader"...
stat output as disabled uploaders

It can happen during the rebalance that the vbucket is not
yet properly instantiated on the node when "uploader" stat
is fetched. The current code treats this as an invalid
output and excludes the bucket from appearing in fusion
status API response.

New code is accepting the discrepancy as a norm and considers
such vbuckets as having the uploader disabled.

Change-Id: I6c28339c465038e8d5aa3dc33b5e0d26f88c7059
Reviewed-on: https://review.couchbase.org/c/ns_server/+/236118
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>