MB-69243: Fix theoretical resource leak in stored_keys.go
When ns_server starts it generates one new integrity token
and saves it in a file (integrity token is basically a string
that is used in integrity checks for encryption key files).
If everything works correctly, old tokens are removed, but if
there is a problem with one of the encryption keys, it is
possible that those tokens are accumulated indefinitely which
is a resource leak (disk).

This change adds a hard limit for the number of those tokens.

Change-Id: I6469e3bf67e0cac7c6c32b2e0d80e9a4270c20b5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/236172
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>