MB-66707: Add --rewrite-key-value parameter to config_remap.erl
Rewrites the value of the given key if found anywhere in ns_config or
chronicle with the provided term. Both key and value are compiled
to the erlang terms according to erlang syntax.

usage:
--rewrite-key-value log_store_uri '"local://tmp"'

Here the key is an atom and the value is a string.

Change-Id: Ie617815678d241d79750cc3d35cd357c242a3c3e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/236420
Tested-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>