MB-66971: Mixed version cluster test
This change provides the infrastructure and tests to for mixed
version and upgrade tests.

This test is significantly different compared with other cluster tests.
The configuration of the cluster is changed as the tests run:

 * only old version nodes,
 * then mixed version
 * then only new version nodes.

During each of the above "phases" a portion (function) from each test
is run. E.g. when "only old version nodes" is present in the cluster
each test runs only their before_upgrade function.

In order to have each test run separately the old version, mixed
version, new version configs would have to be redone for each one.
Instead a single cluster progresses through the phases and the
appropriate function of each test is run.

Initiallly a cluster with two nodes running an older version is started.
And then...

* Each test's before_upgrade function is run.  This allows
  checking/configuring things while on the old version. For example
  these are run at this point (assuming all the upgrade tests are run):

  AlertsUpgradeChecks's before_upgrade function
  BucketSettingsUpgradeChecks's before_upgrade function
  RbacRolesUpgradeChecks's before_upgrade function
  RbacRoleChangesUpgradeChecks's before_upgrade function
  IndexSettingsUpgradeChecks's before_upgrade function

* The nodes running the new version are rebalanced in forming a mixed
  versions cluster.

* Each test's mixed_cluster_check is run to check things while running
  mixed versions. For example, info is obtained via REST from and old
  node and new node...and must match.

* The nodes running the older version are rebalanced out leaving just
  an upgraded cluster running the newer release.

* Each test's post_upgrade_checks function is run.

The cluster is started by specifying run.py with a new arguments
specifying the older version number and the path to ns_server
supporting the older version.

  run.py --older-version 7.6 --older-version-path ~/trinity/ns_server \
         --tests UpgradeChecks

  run.py --older-version 8.0 --older-version-path ~/morpheus/ns_server \
         --tests UpgradeChecks

The tests can be triggered in a gerrit review via:

* To run the upgrade tests for both trinity and morpheus
  make upgrade-tests

* To run the upgrade test from trinity
  make upgrade-from-trinity-tests

* To run the upgrade test from morpheus
  make upgrade-from-morpheus-tests

Change-Id: If21fb48c2ebf17fb87e4f3ee47d8391ce39a95d9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/236530
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>